requirements.txt
backcall==0.2.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dataclasses-json==0.5.2
decorator==4.4.2
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.18.0
jupyter-client==6.1.7
jupyter-core==4.7.0
marshmallow==3.10.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.1.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.3
python-dateutil==2.8.1
python-youtube==0.6.4
pytz==2020.5
pywin32==300
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
stringcase==1.2.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
